Brief citation: remove tokens for links in MARC 856, 952 fields
Tokens intended to allow direct link(s) to online materials from search results. In practice, tokens link to URL in the first 856 field in a record. For serials, this is almost never the 'best' choice; for books, it may link to related, non-full-text eresources, and users many not realize a proper full-text link is available from the full record.